Merge branch 'master' into mzhou/T8646-palette-docker

+++ plugins/tools/basictools/kis_tool_multihand.h
@@ -57,7 +57,7 @@ private Q_SLOTS:
     void activateAxesPointModeSetup();
     void resetAxes();
     void slotSetHandsCount(int count);
-    void slotSetAxesAngle(int angle);
+    void slotSetAxesAngle(qreal angle);
     void slotSetTransformMode(int qcomboboxIndex);
     void slotSetAxesVisible(bool vis);
     void slotSetMirrorVertically(bool mirror);
